# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NYC311 Pulse API",
    "version": "1.0.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Health",
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Healthz Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/meta": {
      "get": {
        "summary": "Meta",
        "operationId": "meta_v1_meta_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Meta V1 Meta Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/snapshot": {
      "get": {
        "summary": "Snapshot",
        "operationId": "snapshot_v1_snapshot_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Snapshot V1 Snapshot Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dimensions": {
      "get": {
        "summary": "Dimensions",
        "operationId": "dimensions_v1_dimensions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Dimensions V1 Dimensions Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/signals": {
      "get": {
        "summary": "Signals",
        "operationId": "signals_v1_signals_get",
        "parameters": [
          {
            "name": "borough",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Borough"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          },
          {
            "name": "problem",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Problem"
            }
          },
          {
            "name": "agency",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agency"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "high",
                    "watch",
                    "research_flag"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Signal"
                  },
                  "title": "Response Signals V1 Signals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/signals/{signal_id}": {
      "get": {
        "summary": "Signal Detail",
        "operationId": "signal_detail_v1_signals__signal_id__get",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Signal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/trends": {
      "get": {
        "summary": "Trends",
        "operationId": "trends_v1_trends_get",
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "const": "volume",
              "type": "string",
              "default": "volume",
              "title": "Metric"
            }
          },
          {
            "name": "signal_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Signal Id"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          },
          {
            "name": "problem",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Problem"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trends V1 Trends Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/map": {
      "get": {
        "summary": "Map Values",
        "operationId": "map_values_v1_map_get",
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "requests",
                "severity"
              ],
              "type": "string",
              "default": "severity",
              "title": "Metric"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Map Values V1 Map Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/quality": {
      "get": {
        "summary": "Quality",
        "operationId": "quality_v1_quality_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Quality V1 Quality Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/evaluation": {
      "get": {
        "summary": "Evaluation",
        "operationId": "evaluation_v1_evaluation_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Evaluation V1 Evaluation Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Evidence": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "label",
          "value",
          "note"
        ],
        "title": "Evidence"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Signal": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "severity": {
            "type": "string",
            "enum": [
              "high",
              "watch",
              "research_flag"
            ],
            "title": "Severity"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "borough": {
            "type": "string",
            "title": "Borough"
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "agency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agency"
          },
          "observed": {
            "type": "number",
            "title": "Observed"
          },
          "expected": {
            "type": "number",
            "title": "Expected"
          },
          "effect": {
            "type": "number",
            "title": "Effect"
          },
          "uncertainty": {
            "type": "string",
            "title": "Uncertainty"
          },
          "persistence": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Persistence"
          },
          "trigger": {
            "type": "string",
            "title": "Trigger"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "data_quality_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data Quality Flags"
          },
          "limitation": {
            "type": "string",
            "title": "Limitation"
          },
          "recommended_action": {
            "type": "string",
            "title": "Recommended Action"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "display_effect": {
            "type": "string",
            "title": "Display Effect"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "episode_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Start"
          },
          "episode_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode End"
          },
          "upper_bound": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upper Bound"
          },
          "excess_count": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excess Count"
          },
          "calibrated_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibrated Score"
          },
          "detector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detector"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "type",
          "severity",
          "as_of",
          "district",
          "borough",
          "problem",
          "observed",
          "expected",
          "effect",
          "uncertainty",
          "persistence",
          "trigger",
          "evidence",
          "data_quality_flags",
          "limitation",
          "recommended_action",
          "title",
          "display_effect"
        ],
        "title": "Signal"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}